CICD

+++ telco_churn/jobs/demo_setup_job.py
@@ -1,3 +1,5 @@
+import os
+
 import mlflow
 from mlflow.tracking import MlflowClient
 from mlflow.exceptions import RestException
@@ -6,7 +8,6 @@ from telco_churn.common import Job
 from telco_churn.utils.logger_utils import get_logger
 
 from databricks.feature_store.client import FeatureStoreClient
-from databricks.feature_store.utils.request_context import RequestContext
 
 client = MlflowClient()
 fs = FeatureStoreClient()
@@ -15,12 +16,12 @@ _logger = get_logger()
 
 class DemoSetup(Job):
 
-    def _check_mlflow_model_registry_exists(self) -> bool:
+    @staticmethod
+    def _check_mlflow_model_registry_exists(model_registry_name) -> bool:
         """
         Check if model exists in MLflow Model Registry.
         Returns True if model exists in Model Registry, False if not
         """
-        model_registry_name = self.conf['mlflow_params']['model_registry_name']
         try:
             client.get_registered_model(name=model_registry_name)
             _logger.info(f'MLflow Model Registry name: {model_registry_name} exists')
@@ -29,32 +30,34 @@ class DemoSetup(Job):
             _logger.info(f'MLflow Model Registry name: {model_registry_name} DOES NOT exists')
             return False
 
-    def _archive_registered_models(self):
+    @staticmethod
+    def _archive_registered_models(model_registry_name):
         """
         Archive any model versions which are not already under stage='Archived
         """
-        registered_model = client.get_registered_model(name=self.conf['mlflow_params']['model_registry_name'])
+        registered_model = client.get_registered_model(name=model_registry_name)
         latest_versions_list = registered_model.latest_versions
 
-        _logger.info(f'MLflow Model Registry name: {self.conf["mlflow_params"]["model_registry_name"]}')
+        _logger.info(f'MLflow Model Registry name: {model_registry_name}')
         for model_version in latest_versions_list:
             if model_version.current_stage != 'Archived':
                 _logger.info(f'Archiving model version: {model_version.version}')
                 client.transition_model_version_stage(
-                    name=self.conf['mlflow_params']['model_registry_name'],
+                    name=model_registry_name,
                     version=model_version.version,
                     stage='Archived'
                 )
 
-    def _delete_registered_model(self):
+    def _delete_registered_model(self, model_registry_name):
         """
         Delete an experiment from the backend store.
         """
-        self._archive_registered_models()
-        client.delete_registered_model(name=self.conf['mlflow_params']['model_registry_name'])
-        _logger.info(f'Deleted MLflow Model Registry model: {self.conf["mlflow_params"]["model_registry_name"]}')
+        self._archive_registered_models(model_registry_name)
+        client.delete_registered_model(name=model_registry_name)
+        _logger.info(f'Deleted MLflow Model Registry model: {model_registry_name}')
 
-    def _check_mlflow_experiments_exists(self) -> dict:
+    @staticmethod
+    def _check_mlflow_experiments_exists() -> dict:
         """
         The demo workflow consists of creating 2 MLflow Tracking experiments:
             * train_experiment - Experiment used to track params, metrics, artifacts during model training
@@ -69,6 +72,11 @@ class DemoSetup(Job):
         -------
         Dictionary indicating whether train and deploy MLflow experiments currently exist
         """
+        train_experiment_id = os.getenv('model_train_experiment_id')    # will be None if not passed
+        train_experiment_path = os.getenv('model_train_experiment_path')
+        deploy_experiment_id = os.getenv('model_deploy_experiment_id')
+        deploy_experiment_path = os.getenv('model_deploy_experiment_path')
+
         def check_by_experiment_id(experiment_id):
             try:
                 mlflow.get_experiment(experiment_id=experiment_id)
@@ -87,26 +95,27 @@ class DemoSetup(Job):
                 _logger.info(f'MLflow Tracking experiment_path: {experiment_path} DOES NOT exist')
                 return False
 
-        mlflow_params = self.conf['mlflow_params']
-
-        if 'train_experiment_id' in mlflow_params:
-            train_exp_exists = check_by_experiment_id(mlflow_params['train_experiment_id'])
-        elif 'train_experiment_path' in mlflow_params:
-            train_exp_exists = check_by_experiment_path(mlflow_params['train_experiment_path'])
+        if train_experiment_id is not None:
+            train_exp_exists = check_by_experiment_id(train_experiment_id)
+        elif train_experiment_path is not None:
+            train_exp_exists = check_by_experiment_path(train_experiment_path)
         else:
-            raise RuntimeError('Either train_experiment_id or train_experiment_path should be passed in demo_setup.yml')
+            raise RuntimeError('Either model_train_experiment_id or model_train_experiment_path should be passed in '
+                               'deployment.yml')
 
-        if 'deploy_experiment_id' in mlflow_params:
-            deploy_exp_exists = check_by_experiment_id(mlflow_params['deploy_experiment_id'])
-        elif 'deploy_experiment_path' in mlflow_params:
-            deploy_exp_exists = check_by_experiment_path(mlflow_params['deploy_experiment_path'])
+        if deploy_experiment_id is not None:
+            deploy_exp_exists = check_by_experiment_id(deploy_experiment_id)
+        elif deploy_experiment_path is not None:
+            deploy_exp_exists = check_by_experiment_path(deploy_experiment_path)
         else:
-            raise RuntimeError('Either train_experiment_id or train_experiment_path should be passed in demo_setup.yml')
+            raise RuntimeError('Either model_train_experiment_id or model_train_experiment_path should be passed in '
+                               'deployment.yml')
 
         return {'train_exp_exists': train_exp_exists,
                 'deploy_exp_exists': deploy_exp_exists}
 
-    def _delete_mlflow_experiments(self, exp_exists_dict: dict):
+    @staticmethod
+    def _delete_mlflow_experiments(exp_exists_dict: dict):
         """
         Check exp_exists_dict if train_exp_exists or deploy_exp_exists is True. Delete experiments if they exist
 
@@ -119,60 +128,78 @@ class DemoSetup(Job):
         if len(delete_experiments) == 0:
             _logger.info(f'No existing experiments to delete')
         if 'train_exp_exists' in delete_experiments:
-            try:
-                mlflow.delete_experiment(experiment_id=self.conf['mlflow_params']['train_experiment_id'])
-                _logger.info(f'Deleted existing experiment_id: {self.conf["mlflow_params"]["train_experiment_id"]}')
-            except KeyError:
-                experiment = mlflow.get_experiment_by_name(name=self.conf['mlflow_params']['train_experiment_path'])
+            if os.getenv('model_train_experiment_path') is not None:
+                experiment = mlflow.get_experiment_by_name(name=os.getenv('model_train_experiment_path'))
                 mlflow.delete_experiment(experiment_id=experiment.experiment_id)
-                _logger.info(f'Deleted existing experiment_path: {self.conf["mlflow_params"]["train_experiment_path"]}')
-        if 'deploy_exp_exists' in delete_experiments:
-            try:
-                mlflow.delete_experiment(experiment_id=self.conf['mlflow_params']['deploy_experiment_id'])
-                _logger.info(f'Deleted existing experiment_id: {self.conf["mlflow_params"]["deploy_experiment_id"]}')
-            except KeyError:
-                experiment = mlflow.get_experiment_by_name(name=self.conf['mlflow_params']['deploy_experiment_path'])
-                mlflow.delete_experiment(experiment_id=experiment.experiment_id)
-                _logger.info(f'Deleted existing experiment_path: {self.conf["mlflow_params"]["deploy_experiment_path"]}')
+                _logger.info(f'Deleted existing experiment_path: {os.getenv("model_train_experiment_path")}')
+            elif os.getenv('model_train_experiment_id') is not None:
+                mlflow.delete_experiment(experiment_id=os.getenv('model_train_experiment_id'))
+                _logger.info(f'Deleted existing experiment_id: {os.getenv("model_train_experiment_id")}')
+            else:
+                raise RuntimeError('Either model_train_experiment_id or model_train_experiment_path should be passed '
+                                   'in deployment.yml')
 
-    def _check_feature_table_exists(self) -> bool:
+        if 'deploy_exp_exists' in delete_experiments:
+            if os.getenv('model_deploy_experiment_path') is not None:
+                experiment = mlflow.get_experiment_by_name(name=os.getenv('model_deploy_experiment_path'))
+                mlflow.delete_experiment(experiment_id=experiment.experiment_id)
+                _logger.info(
+                    f'Deleted existing experiment_path: {os.getenv("model_deploy_experiment_path")}')
+            elif os.getenv('model_deploy_experiment_id') is not None:
+                mlflow.delete_experiment(experiment_id=os.getenv('model_deploy_experiment_id'))
+                _logger.info(f'Deleted existing experiment_id: {os.getenv("model_deploy_experiment_id")}')
+
+    @staticmethod
+    def _check_feature_table_exists(feature_store_table) -> bool:
         """
         Check if Feature Store feature table exists
         Returns True if feature table exists in Feature Store, False if not
         """
         try:
-            fs.get_table(name=self.conf['feature_store_table'])
-            _logger.info(f'Feature Store feature table: {self.conf["feature_store_table"]} exists')
+            fs.get_table(name=feature_store_table)
+            _logger.info(f'Feature Store feature table: {feature_store_table} exists')
             return True
         except (ValueError, Exception):
-            _logger.info(f'Feature Store feature table: {self.conf["feature_store_table"]} DOES NOT exist')
+            _logger.info(f'Feature Store feature table: {feature_store_table} DOES NOT exist')
             return False
 
-    def _delete_feature_table(self):
+    @staticmethod
+    def _drop_feature_table(feature_store_table):
         """
         Delete Feature Store feature table
-
-        TODO: UPDATE this method once the public API to delete a feature table has been released. Currently uses protected class
         """
-        ####################################################################
-        # DELETE once public API to delete feature table
-        rq = RequestContext(feature_store_method_name='test_only_method')
-        fs._catalog_client.delete_feature_table(self.conf['feature_store_table'], req_context=rq)
-        _logger.info(f'Deleted Feature Store feature table: {self.conf["feature_store_table"]}')
-        ####################################################################
-
-    def _check_labels_delta_table_exists(self) -> bool:
         try:
-            _logger.info(f'Labels Delta table: {self.conf["labels_table_path"]} exists')
-            self.dbutils.fs.ls(self.conf['labels_table_path'])
+            fs.drop_table(
+                name=feature_store_table
+            )
+            _logger.info(f'Deleted Feature Store feature table: {feature_store_table}')
+        except ValueError:
+            _logger.info(f'Feature Store feature table: {feature_store_table} does not exist')
+
+    def _check_labels_delta_table_exists(self, labels_table_dbfs_path) -> bool:
+        """
+        Check if Delta table exists in DBFS
+
+        Parameters
+        ----------
+        labels_table_dbfs_path : str
+            Path to Delta table in DBFS
+
+        Returns
+        -------
+        bool
+        """
+        try:
+            self.dbutils.fs.ls(labels_table_dbfs_path)
+            _logger.info(f'Labels Delta table: {labels_table_dbfs_path} exists')
             return True
         except:
-            _logger.info(f'Labels Delta table: {self.conf["labels_table_path"]} does not exist')
+            _logger.info(f'Labels Delta table: {labels_table_dbfs_path} DOES NOT exist')
             return False
 
-    def _delete_labels_delta_table(self):
-        self.dbutils.fs.rm(self.conf['labels_table_path'], True)
-        _logger.info(f'Deleted labels Delta table: {self.conf["labels_table_path"]}')
+    def _delete_labels_delta_table(self, labels_table_dbfs_path):
+        self.dbutils.fs.rm(labels_table_dbfs_path, True)
+        _logger.info(f'Deleted labels Delta table: {labels_table_dbfs_path}')
 
     def setup(self):
         """
@@ -182,22 +209,32 @@ class DemoSetup(Job):
         * Delete Feature Table if exists
         """
         _logger.info('==========Demo Setup=========')
+        _logger.info(f'Running demo-setup pipeline in {os.getenv("DEPLOYMENT_ENV")} environment')
 
-        _logger.info('Checking MLflow Model Registry...')
-        if self._check_mlflow_model_registry_exists():
-            self._delete_registered_model()
+        if self.conf['delete_model_registry']:
+            _logger.info('Checking MLflow Model Registry...')
+            model_registry_name = os.getenv('model_registry_name')
+            if self._check_mlflow_model_registry_exists(model_registry_name):
+                self._delete_registered_model(model_registry_name)
 
-        _logger.info('Checking MLflow Tracking...')
-        exp_exists_dict = self._check_mlflow_experiments_exists()
-        self._delete_mlflow_experiments(exp_exists_dict)
+        if self.conf['delete_mlflow_experiments']:
+            _logger.info('Checking MLflow Tracking...')
+            exp_exists_dict = self._check_mlflow_experiments_exists()
+            self._delete_mlflow_experiments(exp_exists_dict)
 
-        _logger.info('Checking Feature Store...')
-        if self._check_feature_table_exists():
-            self._delete_feature_table()
+        if self.conf['drop_feature_table']:
+            _logger.info('Checking Feature Store...')
+            feature_store_database_name = os.getenv('feature_store_database_name')
+            feature_store_table_name = os.getenv('feature_store_table_name')
+            feature_store_table = f'{feature_store_database_name}.{feature_store_table_name}'
+            if self._check_feature_table_exists(feature_store_table=feature_store_table):
+                self._drop_feature_table(feature_store_table=feature_store_table)
 
-        _logger.info('Checking existing labels table...')
-        if self._check_labels_delta_table_exists():
-            self._delete_labels_delta_table()
+        if self.conf['drop_labels_table']:
+            _logger.info('Checking existing labels table...')
+            labels_table_dbfs_path = os.getenv('labels_table_dbfs_path')
+            if self._check_labels_delta_table_exists(labels_table_dbfs_path=labels_table_dbfs_path):
+                self._delete_labels_delta_table(labels_table_dbfs_path=labels_table_dbfs_path)
 
         _logger.info('==========Demo Setup Complete=========')
 
@@ -205,9 +242,9 @@ class DemoSetup(Job):
         """
         Launch DemoSetup job
         """
-        _logger.info("Launching DemoSetup job")
+        _logger.info('Launching DemoSetup job')
         DemoSetup().setup()
-        _logger.info("DemoSetup job finished!")
+        _logger.info('DemoSetup job finished!')
 
 
 if __name__ == "__main__":
